requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
pycryptodome
python-multipart
pytest